+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-- [TD-7011] Format.search_values/2 doesn't apply default values by default
+- [TD-7011] Format.search_values/2 doesn't apply default values when specified
 
 ## [6.15.0] 2024-11-27